refactor: 모델에 card 패키지 추가, 파일 이동
