+++ src/main/kotlin/blackjack/model/Participant.kt
@@ -1,5 +1,7 @@
 package blackjack.model
 
+import blackjack.model.card.Card
+
 abstract class Participant(val name: String) {
     private val _cards: MutableList<Card> = mutableListOf()
     val cards get() = _cards.toList()
