+++ src/main/kotlin/blackjack/model/card/CardNumber.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.card
 
 enum class CardNumber(val score: Int) {
     ACE(11),
