+++ src/main/kotlin/blackjack/view/OutputView.kt
@@ -1,18 +1,18 @@
 package blackjack.view
 
-import blackjack.model.Card
+import blackjack.model.card.Card
 import blackjack.model.Dealer
 import blackjack.model.GameManager.Companion.INITIAL_HAND_OUT_CARD_COUNT
-import blackjack.model.CardNumber
-import blackjack.model.CardNumber.ACE
-import blackjack.model.CardNumber.JACK
-import blackjack.model.CardNumber.KING
-import blackjack.model.CardNumber.QUEEN
+import blackjack.model.card.CardNumber
+import blackjack.model.card.CardNumber.ACE
+import blackjack.model.card.CardNumber.JACK
+import blackjack.model.card.CardNumber.KING
+import blackjack.model.card.CardNumber.QUEEN
 import blackjack.model.Player
 import blackjack.model.ResultCalculator.adjustScore
 import blackjack.model.ResultCalculator.calculateTotalScore
 import blackjack.model.ResultType
-import blackjack.model.Shape
+import blackjack.model.card.Shape
 
 class OutputView {
     fun printInitialHandOutCardMessage(players: List<Player>) {
