+++ src/test/kotlin/blackjack/DealerTest.kt
@@ -1,9 +1,9 @@
 package blackjack
 
-import blackjack.model.Card
+import blackjack.model.card.Card
 import blackjack.model.Dealer
-import blackjack.model.CardNumber
-import blackjack.model.Shape
+import blackjack.model.card.CardNumber
+import blackjack.model.card.Shape
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
